feat(order): added create order, update order, get order by id

+++ order-service/internal/data/models/restaurant_order.go
@@ -3,11 +3,11 @@ package models
 import "time"
 
 type RestaurantOrder struct {
-	ID         int        `db:"id"`
-	CreatedAt  time.Time  `db:"created_at"`
-	UpdatedAt  time.Time  `db:"updated_at"`
-	DeletedAt  *time.Time `db:"deleted_at"`
-	MenuIDs    []int      `db:"menu_ids"`
-	Total      float64    `db:"total"`
-	Status     string     `db:"status"`
+	ID        int        `db:"id"`
+	CreatedAt time.Time  `db:"created_at"`
+	UpdatedAt time.Time  `db:"updated_at"`
+	DeletedAt *time.Time `db:"deleted_at"`
+	MenuIDs   []int      `db:"menu_ids"`
+	Total     float32    `db:"total"`
+	Status    string     `db:"status"`
 }
